2927: Renamed output models
